Updated prebuilt design
